requirements.txt
grpcio==1.67.1
google-generativeai
pdf2image
pillow
PyPDF2
streamlit